% Generated by roxygen2 (4.0.1): do not edit by hand
\name{foo}
\alias{foo}
\title{Test namespace}
\usage{
foo(DF, index = list(1:5, 6:10))
}
\arguments{
\item{DF}{A DataFrame}

\item{index}{A list of length equal to the number of columns in the DataFrame
where each element is an index for that column.}
}
\value{
A subset of the data
}
\description{
Test namespace correctness
}
\examples{
## Initial info
sessionInfo()

## Load data
load("~/Desktop/DF.Rdata")

## Run function
result <- foo(DF)
}

